% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_geo_headers}
\alias{clean_geo_headers}
\title{Normalize messy geography headers to standard names}
\usage{
clean_geo_headers(df, map, keep)
}
\arguments{
\item{df}{A data frame with geography columns.}

\item{map}{Named character vector: names are standardized outputs,
values are regex patterns of accepted input names.}

\item{keep}{Character vector of standardized columns to keep.}
}
\value{
A tibble with standardized names.
}
\description{
Normalize messy geography headers to standard names
}
